# src/swagger/paths.yaml
openapi: 3.0.3
info:
  title: Swagger Petstore - OpenAPI 3.0
  description: |-
    This is a sample Pet Store Server based on the OpenAPI 3.0 specification.  You can find out more about
    Swagger at [https://swagger.io](https://swagger.io). In the third iteration of the pet store, we've switched to the design first approach!
    You can now help us improve the API whether it's by making changes to the definition itself or to the code.
    That way, with time, we can improve the API in general, and expose some of the new features in OAS3.

    _If you're looking for the Swagger 2.0/OAS 2.0 version of Petstore, then click [here](https://editor.swagger.io/?url=https://petstore.swagger.io/v2/swagger.yaml). Alternatively, you can load via the `Edit > Load Petstore OAS 2.0` menu option!_

    Some useful links:
    - [The Pet Store repository](https://github.com/swagger-api/swagger-petstore)
    - [The source API definition for the Pet Store](https://github.com/swagger-api/swagger-petstore/blob/master/src/main/resources/openapi.yaml)
  termsOfService: http://swagger.io/terms/
  contact:
    email: apiteam@swagger.io
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.11
...
servers:
  - url: http://localhost:3001/
...
paths:
...
  /api/recruiter/sign-in-recruiter:
    post:
      tags:
        - recruiter
      summary: Đăng nhập nhà tuyển dụng
      operationId: loginRecruiter
      requestBody:
        description: "Đăng nhập nhà tuyển đụng"
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
        required: true
      responses:
        "200":
          description: Đăng nhập thành công
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessLogin"
        "422":
          description: Invalid input
...
  /api/recruiter/delete-recruiter/{id}:
    delete:
      tags:
        - recruiter
      summary: Xóa tài khoản nhà tuyển dụng.
      operationId: deleteRecruiter
      parameters:
        - name: id
          in: path
          description: Xóa tài khoản nhà tuyển dụng.
          required: true
          schema:
            type: string
            format: MongoId
            example: "66dcb19ef533c72b0b0b94d4"
      responses:
        "200":
          description: Xóa tài khoản thành công
        "400":
          description: Invalid input
        "404":
          description: User not found
  /api/recruiter/change-password-recruiter:
    put:
      tags:
        - recruiter
      summary: Đổi mật khẩu nhà tuyển dụng
      operationId: changePasswordRecruiter
      requestBody:
        description: "Đổi mật khẩu nhà tuyển dụng"
        content:
          application/json:
            schema:
              properties:
                userId:
                  type: string
                oldPassword:
                  type: string
                newPassword:
                  type: string
        required: true
      responses:
        "200":
          description: Đổi mật khẩu thành công
        "422":
          description: Invalid input
  /api/recruiter/update-recruiter/{id}:
    put:
      tags:
        - recruiter
      summary: Cập nhập thông tin nhà tuyển dụng
      operationId: updateRecruiter
      parameters:
        - name: id
          in: path
          description: Cập nhập thông tin nhà tuyển dụng
          required: true
          schema:
            type: string
            format: MongoId
            example: "66dcb19ef533c72b0b0b94d4"
      requestBody:
        content:
          application/json:
            schema:
              properties:
                fullName:
                  type: string
                  example: Nguyen Van A
                companyName:
                  type: string
                  example: Ctu company
                companyWebsite:
                  type: string
                  example: https://ctu.edu.vn
                companyFacebook:
                  type: string
                  example: https://www.facebook.com/CTUDHCT/?locale=vi_VN
                phoneNumber:
                  type: number
                  example: 0123456789
                companyDescription:
                  type: string
                  example: Thông tin công ty
                companyAddress:
                  type: string
                  example: Khu II, Đ. 3 Tháng 2, Xuân Khánh, Ninh Kiều, Cần Thơ
                businessLicense:
                  type: string
                  example: "http://image.html"
      responses:
        "200":
          description: Cập nhật thông tin thành công
        "422":
          description: Invalid input
  /api/recruiter/getAll-recruiter:
    get:
      tags:
        - recruiter
      summary: Lấy tất cả tài khoản nhà tuyển dụng
      operationId: getAllRecruiter
      responses:
        "200":
          description: Lấy thông tin người dùng thành công
        "422":
          description: Invalid input
...
  /api/recruiter/delete-many-recruiter:
    post:
      tags:
        - recruiter
      summary: Xóa nhiều tài khoản nhà tuyển dụng cùng lúc
      operationId: deleteManyRecruiter
      requestBody:
        description: "Xóa nhiều tài khoản nhà tuyển dụng cùng lúc"
        content:
          application/json:
            schema:
              properties:
                userId:
                  type: array
                  example:
                    [
                      "66dcb19ef533c72b0b0b94d4",
                      "66dcb19ef533c72b0b0b94d5",
                      "66dcb19ef533c72b0b0b94d6",
                    ]
        required: true
      responses:
        "200":
          description: Xóa người dùng thành công
        "422":
          description: Invalid input
...
components:
  schemas:
    LoginBody:
      type: object
      properties:
        email:
          type: string
          example: nguyenvana@gmail.com
        password:
          type: string
          example: nguyenvana123
    SuccessLogin:
      type: object
      properties:
        status:
          type: string
          example: OK
        message:
          type: string
          example: Success
        access_token:
          type: string
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6IjY2ZDg5YTRhZjVkZTJlNTBlNjNkZmQzZiIsImlhdCI6MTcyNTU2Mzk3MiwiZXhwIjoxNzI1NTY0MDAyfQ.h-_VdUIvxPz57WkUZWSc7Z4z6XJufzR7HODdqJnUBx0
        refresh_token:
          type: string
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6IjY2ZDg5YTRhZjVkZTJlNTBlNjNkZmQzZiIsImlhdCI6MTcyNTU2Mzk3MiwiZXhwIjoxNzU3MDk5OTcyfQ.Gh8RxZrjYIp1GQjTkv19JkHWd7SPQAsDyBdaGViTXx
...